remove es client import from reindex script, to avoid crashes in environments not configured with ELASTICSEARCH_URL
